refactor(utils): move blake to eddsa-poseidon and make buffer dev dependency
