+++ packages/utils/src/index.ts
@@ -12,4 +12,3 @@ export * from "./conversions"
 export * from "./proof-packing"
 export * from "./type-checks"
 export * from "./error-handlers"
-export * from "./blake512"
